add zero3 param manager
